Split index.js into many smaller files
